backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
redis==5.0.1
sentence-transformers==3.3.1
pypdf==3.17.0
faiss-cpu
openai==1.3.6
python-multipart==0.0.6
python-dotenv==1.0.0
huggingface-hub
numpy==1.26.0
google-generativeai==0.3.2
protobuf>=3.19.5,<5.0.0dev
